Fix the Huge engine, CLI, spatial/features/vocabulary and vendored-zstd findings
Huge engine (BG-289, BG-415, BG-355): the disk engines took the entity prefix of the sorted object column up to the first LITERAL, and triple terms sort after literals, so a store with no top-level literal object streamed its triple terms into the entities dictionary, whose writer threw; the prefix now ends at the first literal or triple term (TripleTermOnlyObjectsTest, all six engines). Every writer published its finished build inside the catch that deletes a failed build's temp file, so a destination that refused the rename (a memory-mapped store on Windows) destroyed a completed store; core.AtomicPublish retries briefly, then KEEPS the build as <dest>.new and names it (AtomicPublishTest; used by all six writers and -export). The disk engines' grid-URN format shares Params.gridGraph with the RAM engine (LocaleIndependentIdentifiersTest compares both under ar-EG).

CLI (BG-145, BG-417, BG-261, BG-143, BG-393, BG-420): Params.isInternalGraph (urn:x-beakgraph:*) replaces the VoID/Spatial-only export filter, so a -spatial store exports as plain triples without its tile graphs; destination mapping resolves paths to absolute so "-src ." works; a single-file -src writes the named -dest (or <dest>/<name>.h5 into a directory); FileProcessor catches Throwable and reports a missing -dest per file, and the writers clean their temp file on Error; -base now also drives -export (document-relative IRIs resolved on both the index and generic paths; NT/NQ without -base fails naming the option, TTL/TriG/JSON-LD warn); destinations are planned first so sources colliding on one .h5 are reported and counted, skipped existing destinations are logged and counted, and every writer builds into a unique temp name. ExportTest and BeakGraphCLIConversionTest cover each; docs/INSTRUCTIONS.md updated.

Spatial, features and vocabularies (BG-146, 151, 69, 70, 74, 373, 163, 149, 377, 150, 148, 160, 152, 154, 364, 305, 306, 153, 161, 379, 307): the polygon raster honours beakgraph.features.maxRasterPixels with pixel-derived features rescaled; MULTIPOLYGON gets shape and axis features; the eigenvalue discriminant uses hypot and clamps (no NaN axes on large near-isotropic regions); geof:sfIntersects prepares and caches the query constant per thread, parses the CRS prefix (an unprefixed literal adopts the other operand's CRS; two explicit differing CRSs are an evaluation error - README/SPECIFICATIONS say no CRS transformation is performed) and is installed in a dataset-scoped FunctionRegistry so BG answers do not depend on what an embedder registers globally (Spatial.init no longer overwrites an existing registration); PolygonScaler works on a copy, repairs invalid rings through fixPolygon with a warning instead of a silent buffer(0), and skips unrepairable geometry without throwing; WKTDatatype registers deterministically (JenaSystem.init first) with a cheap syntactic isValid; LWS.java keeps only vocabulary terms (plus Storage/StorageRoot) and the servlet types the description lws:DataResource; HilbertSpace's viewer helpers use floor+clamp, refuse off-domain neighbours and bound the boundary walk (beakgraph.hilbert.maxWalkCells) - kept as Halcyon API rather than deleted; GridCell keys are ASCII; Box, EXIF, Vector2D and the unused ShapeAnalysis/ImageTools/Resource-overload members are deleted; GEO drops its unused Model. SpatialFeatureFixesTest and IntersectsFunctionTest.

Vendored zstd (BG-166, 170, 174, 172, 173, 168): the Huffman table is cleared per frame (a treeless-literals block opening a frame is rejected instead of decoding against the previous fragment's table); corrupted Huffman/FSE headers raise MalformedInputException instead of array exceptions (int symbol loops, verified weights and counts); three copy-pasted verify messages corrected; the unreachable streaming/Hadoop classes deleted; the README pins the upstream commit, lists every local divergence and a re-vendor recipe (the fork commit was not recorded and says so). ZstdCodecInteropTest: round trips, native<->Java interop (Linux CI), bit-flip and truncation fuzzing, the treeless case and the block-size message. BG-171 (compressor context reuse) is deliberately left open.

Full suite: 1797 tests, 0 failures, 91 skipped (the extra skip is the native zstd interop case on a machine without the native library).

+++ src/main/java/io/airlift/compress/v3/zstdFFM/FseCompressionTable.java
@@ -144,9 +144,9 @@ class FseCompressionTable
         int step = calculateStep(tableSize);
 
         int position = 0;
-        for (byte symbol = 0; symbol <= maxSymbolValue; symbol++) {
+        for (int symbol = 0; symbol <= maxSymbolValue; symbol++) {
             for (int i = 0; i < normalizedCounters[symbol]; i++) {
-                symbols[position] = symbol;
+                symbols[position] = (byte) symbol;
                 do {
                     position = (position + step) & mask;
                 }
